hf-space-deploy/requirements_full.txt
# Core requirements
gradio==4.16.0
opencv-python-headless==4.8.1.78
numpy==1.24.3
websockets==12.0
feedparser==6.0.10
requests==2.31.0
Pillow==10.1.0

# AI Models for REAL character generation and animation
torch==2.1.0
torchvision==0.16.0
transformers==4.35.0
diffusers==0.24.0
accelerate==0.24.0
safetensors==0.4.0
xformers==0.0.22

# Character Animation & Lip Sync
# Wav2Lip
filelock==3.12.2
numba==0.57.1
librosa==0.10.1
opencv-contrib-python==4.8.1.78

# TTS - Bark is available on HF and works well
bark==1.0.0
# Alternative: TTS package includes multiple models
TTS==0.21.0

# Face Enhancement
gfpgan==1.3.8
facexlib==0.3.0
realesrgan==0.3.0

# Video Processing
moviepy==1.0.3
imageio==2.31.1
imageio-ffmpeg==0.4.8

# Additional utilities
scipy==1.11.4
scikit-image==0.21.0
tqdm==4.66.1

# For YouTube/video sourcing (optional)
yt-dlp==2023.11.16
pytube==15.0.0

# Audio processing
soundfile==0.12.1
pydub==0.25.1
simpleaudio==1.0.4